% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-categoric-dists.R
\name{categoric_dists}
\alias{categoric_dists}
\title{Visualise distributions of categorical columns in a dataframe}
\usage{
categoric_dists(
  x,
  cols = dplyr::everything(),
  top_n = 5,
  na_colour = "blue",
  other_colour = "grey50"
)
}
\arguments{
\item{x}{A dataframe containing character columns}

\item{cols}{Tidyselect columns to use (default = \code{dplyr::everything()})}

\item{top_n}{The top n levels to show - all other values will be lumped into 'Other'}

\item{na_colour}{Colour for NA bars}

\item{other_colour}{Colour for lumped 'Other' bars}
}
\description{
Returns a facetted plot of the count and percent of the top levels in each categorical column
of a dataframe
}
