% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-geom-origin.R
\name{geom_origin}
\alias{geom_origin}
\title{Add vertical and horizontal lines through the origin}
\usage{
geom_origin(...)
}
\arguments{
\item{...}{Arguments passed to \code{ggplot2::geom_segment()}}
}
\description{
Add vertical and horizontal lines through the origin
}
\examples{
mtcars \%>\%
ggplot(aes(wt, disp))+
  geom_point(aes(col = mpg))+
  geom_origin()
}
