% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-move-legend.R
\name{move_legend}
\alias{move_legend}
\title{Move (or remove) a ggplot2 legend}
\usage{
move_legend(where = "none")
}
\arguments{
\item{where}{One of "none", "left", "right", "bottom", "top", or two-element numeric vector
(default = "none")}
}
\value{

}
\description{
A convenience wrapper to move a \code{ggplot2} legend
}
\examples{
mtcars \%>\%
ggplot(aes(wt, disp))+
  geom_point(aes(col = mpg))+
  move_legend("left")
}
