% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-numeric-dists.R
\name{numeric_dists}
\alias{numeric_dists}
\title{Visualise distributions of numeric columns in a dataframe}
\usage{
numeric_dists(
  x,
  cols = dplyr::everything(),
  remove_outliers = FALSE,
  bp_thickness = 0.12,
  bp_center = 0,
  density_colour = "grey",
  bp_colour = "blue",
  ecdf_colour = "red"
)
}
\arguments{
\item{x}{dataframe containing numeric columns}

\item{cols}{Tidyselect columns to use (default = \code{dplyr::everything()})}

\item{remove_outliers}{Rremove outliers before visualisation (default = FALSE)}

\item{bp_thickness}{Boxplot thickness (in units of y)}

\item{bp_center}{Boxplot centre height (in units of y)}

\item{density_colour}{Density estimate fill colour}

\item{bp_colour}{Boxplot colour}

\item{ecdf_colour}{ECDF colour}
}
\description{
Returns a facetted plot of the density estimate (scaled to 1) including an
overlay of the ECDF and horizontal box plot
}
\examples{
numeric_dists(diamonds[1:3000, ])
}
