% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-print-all.R
\name{print_all}
\alias{print_all}
\title{Print all rows of a dataframe to the console}
\usage{
print_all(x, n = nrow(x))
}
\arguments{
\item{x}{A dataframe}

\item{n}{Number of rows to print (default = \code{nrow(x)})}
}
\description{
Risky! Use with caution
}
