% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat-remove-outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers from a vector of numbers}
\usage{
remove_outliers(x, na_fill = TRUE, ...)
}
\arguments{
\item{x}{A vector of numbers}

\item{na_fill}{Replace outliers with NA, keeping the length of x consistent (default = TRUE)}

\item{...}{Further arguments passed to \code{grDevices::boxplot.stats()}}
}
\description{
Remove outliers from a vector of numbers
}
