% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-rotate-x-labs.R
\name{rotate_x_labs}
\alias{rotate_x_labs}
\title{Rotate ggplot2 x axis text}
\usage{
rotate_x_labs(angle = 90, hjust = 1, vjust = 0.5, ...)
}
\arguments{
\item{angle}{Angle of text rotation (default = 90)}

\item{hjust}{Horizontal text justification (default = 1)}

\item{vjust}{Vertical text justification (default = 0.5)}

\item{...}{Further arguments passed to \code{ggplot2::element_text()}}
}
\description{
A convenience function to quickly rotate x-axis labels
}
\examples{
mtcars \%>\%
ggplot(aes(wt, disp))+
geom_point()+
rotate_x_labs()
}
