% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-theme-holmes.R
\name{theme_holmes}
\alias{theme_holmes}
\title{My ggplot2 plotting theme}
\usage{
theme_holmes(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{base font size given in pts}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}
}
\description{
Some small tweaks to \code{ggplot2::theme_bw()}
}
